sistema/requirements.txt
asgiref==3.5.2
certifi==2022.9.24
charset-normalizer==2.1.1
Django==4.1.2
future==0.18.2
idna==3.4
iso8601==1.1.0
pyserial==3.5
PyYAML==6.0
requests==2.28.1
sqlparse==0.4.3
urllib3==1.26.12
